Adjusted settings and codes
Removed unnecessary codes and ladderlog events from the script. This
includes NEW_MATCH since it can also be triggered when only one player
is active and suddenly a new player comes on. That is when a new match
is created and will break the purpose of the script.

Added the counting of humans, ais, humans alive, ais alive, humans dead,
ais dead. Very handy for the activation of the timer.

Modified the timer activation when there are actually dead players
active in the server. In the future this will need improvement since it
won't activate when a new player enters the server. Also added the
COLLAPSE_ALL to collapse all zones when there are no players alive.