Stablised the settings and script
Fixed up the many mistakes, errors and other factors adding into causing
the script to not function its operations properly.

Using script powered round rotation for maps which is also easier since
when the count reaches max is when the race will finish. So, all players
would need to do is hit "/start" and start racing to 100 maps/rounds!

The script is also flexible for allowing players to pause/resume the
tournament progress. Simply by executing the chat commands "/pause" to
pause the race and "/resume" from where they left of.